Avoid signed/unsigned warnings in FunctionDAG.h

+++ apps/autoscheduler/FunctionDAG.h
@@ -133,13 +133,13 @@ public:
             // The producer is scalar, so all strides are zero.
             return {true, 0, 1};
         }
-        internal_assert(producer_storage_dim < coeffs.size());
+        internal_assert(producer_storage_dim < (int) coeffs.size());
         const auto &p = coeffs[producer_storage_dim];
         if (p.empty()) {
             // The consumer is scalar, so all strides are zero.
             return {true, 0, 1};
         }
-        internal_assert(consumer_loop_dim < p.size());
+        internal_assert(consumer_loop_dim < (int) p.size());
         return p[consumer_loop_dim];
     }
